Fixed daily jobs, model properties name + added api_service folder

+++ webpeditor_app/services/image_services/image_service.py
@@ -9,48 +9,45 @@ import requests
 from PIL import Image as PilImage, ExifTags
 from PIL.Image import Image as ImageClass
 from _decimal import ROUND_UP, Decimal
-from django.contrib.sessions.backends.db import SessionStore
 from django.http import JsonResponse
 from rest_framework.utils.serializer_helpers import ReturnDict
 
 from webpeditor_app.models.database.models import OriginalImage, EditedImage
 from webpeditor_app.models.database.serializers import OriginalImageSerializer, EditedImageSerializer
+from webpeditor_app.services.api_services.cloudinary_service import delete_user_folder_with_content
 
 logging.basicConfig(level=logging.INFO)
 
 
-def delete_old_image_in_db_and_local(user_id: str) -> JsonResponse:
-    """
-    Deletes the expired image session and the corresponding image from the user's media folder.
+def delete_old_original_and_edited_image(user_id: str) -> JsonResponse:
+    delete_original_image_in_db(user_id)
+    delete_edited_image_in_db(user_id)
 
-    Args:
-        user_id (str): The ID of the user whose image and image session is being checked.
-
-    Returns:
-        A JSON response with success status and information message.
-    """
-    original_image = get_original_image(user_id)
-    if original_image:
-        original_image.delete()
-        logging.info("Original image has been deleted from db.")
-
-        session_store = SessionStore(session_key=original_image.session_id)
-        session_store.clear_expired()
-        logging.info("Session has been cleared")
-
-    edited_image = get_edited_image(user_id)
-    if edited_image:
-        edited_image.delete()
-        logging.info("Edited image has been deleted from db.")
-
-    # Delete user's folder in all other cases
-    if original_image is None or edited_image is None:
-        logging.info("No images in db. Deleting user folder...")
+    delete_user_folder_with_content(user_id)
 
     return JsonResponse({
         'success': True,
         'info': 'Session has been expired and image has been deleted'
     }, status=204)
+
+
+def delete_original_image_in_db(user_id: str):
+    original_image = get_original_image(user_id)
+    if original_image is None:
+        logging.info("No original image in db. Deleting user's folder...")
+        delete_user_folder_with_content(user_id)
+
+    original_image.delete()
+    logging.info("Original image has been deleted from db")
+
+
+def delete_edited_image_in_db(user_id: str):
+    original_image = get_original_image(user_id)
+    edited_image = get_edited_image(user_id)
+    # Check if original image was deleted
+    if original_image is None and edited_image is not None:
+        edited_image.delete()
+        logging.info("Edited image has been deleted from db")
 
 
 def get_serialized_data_original_image() -> ReturnDict:
@@ -98,7 +95,6 @@ def data_url_to_binary(data_url: str) -> BytesIO:
     return BytesIO(image_data)
 
 
-
 def get_edited_image(user_id: str) -> EditedImage | None:
     edited_image = EditedImage.objects.filter(user_id=user_id).first()
     if edited_image is None:
@@ -107,9 +103,10 @@ def get_edited_image(user_id: str) -> EditedImage | None:
     return edited_image
 
 
-def get_image_file_instance(image_file_bytes: BytesIO) -> ImageClass | None:
+def get_image_file_instance(image_data: BytesIO) -> ImageClass | None:
     try:
-        return PilImage.open(image_file_bytes)
+        image_data.seek(0)
+        return PilImage.open(image_data)
     except Exception as e:
         logging.error(e)
         return None
@@ -173,10 +170,9 @@ def get_image_file_size(image: ImageClass) -> str:
         return f"{size_in_kb:.1f} KB"
 
 
-def get_info_about_image(image_file_bytes: BytesIO) \
+def get_info_about_image(image_data: BytesIO) \
         -> None | Tuple:
-    image_file = get_image_file_instance(image_file_bytes)
-
+    image_file = get_image_file_instance(image_data)
     if image_file is None:
         return None
 
@@ -185,21 +181,19 @@ def get_info_about_image(image_file_bytes: BytesIO) \
     image_resolution = f"{image_file.width}px ⨯ {image_file.height}px"
     image_aspect_ratio = get_image_aspect_ratio(image_file)
     image_mode = image_file.mode
-    image_format = str(image_file.format).lower()
+    image_format = image_file.format
+    metadata = image_file.info
 
     # Get structured exif data, if exists
     exif_data = image_file.getexif()
     if len(exif_data) == 0:
         exif_data = "No exif data was found"
     else:
-        exif = {
+        exif_data = {
             ExifTags.TAGS[k]: v
             for k, v in exif_data.items()
             if k in ExifTags.TAGS
         }
-        exif_data = exif
-
-    metadata = image_file.info
 
     image_file.close()
